feat:successfully fetching and showing subject test results per subjects

+++ lib/src/features/dashboard/api/school/models/fetch_test_exams_response/tests/fetch_test_by_subject/fetch_subject_test_result_grade.dart
@@ -12,7 +12,7 @@ class FetchSubjectTestResultGrade {
   factory FetchSubjectTestResultGrade.fromJson(Map<String, dynamic> json) =>
       _$FetchSubjectTestResultGradeFromJson(json);
   @JsonKey(name: 'grade')
-  final int grade;
+  final double grade;
 
   @JsonKey(name: '_id')
   final String id;
